local_python/pyproject.toml
[project]
name = "local-python"
version = "0.1.0"
description = "Chess.com Blunder Trainer - analyze games and create training puzzles"
authors = [
    {name = "Your Name", email = "<your-email>"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "python-chess>=1.10.0",
    "pytest>=8.0.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
